wip type safety
