+++ .eslintrc.json
@@ -82,6 +82,8 @@
       "files": ["**/*.{ts,tsx}"],
       "extends": ["plugin:@typescript-eslint/recommended"],
       "rules": {
+        "no-shadow": 0,
+        "@typescript-eslint/no-shadow": 2,
         "@typescript-eslint/no-explicit-any": 2,
         "@typescript-eslint/no-unnecessary-condition": 2,
         "@typescript-eslint/explicit-module-boundary-types": 0,
